+++ product_common/src/error.rs
@@ -1,9 +1,6 @@
 // Copyright 2020-2025 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-#[cfg(feature = "gas-station")]
-use crate::transaction::transaction_builder::gas_station::GasStationError;
-
 /// Alias for a `Result` with the error type [`Error`].
 pub type Result<T, E = Error> = core::result::Result<T, E>;
 
@@ -52,5 +49,5 @@ pub enum Error {
   /// Gas Station specific error.
   #[cfg(feature = "gas-station")]
   #[error("gas-station sponsoring failed: {0}")]
-  GasStation(#[from] GasStationError),
+  GasStation(#[from] crate::gas_station::GasStationError),
 }
