+++ product_common/src/lib.rs
@@ -6,6 +6,8 @@ pub mod bindings;
 #[cfg(feature = "core-client")]
 pub mod core_client;
 pub mod error;
+#[cfg(feature = "gas-station")]
+pub mod gas_station;
 #[cfg(feature = "http-client")]
 pub mod http_client;
 pub mod network_name;
